Refactored Challenge 3
