+++ student.rb
@@ -1,15 +1,14 @@
-class Student
-  attr_reader :age, :phase
-  attr_accessor :name
+require_relative 'person'
+require_relative 'highfiveable'
 
+class Student < Person
+
+  include Highfiveable
+  attr_reader :phase
+  
   def initialize(options = {})
+    super
     @phase = 1
-    @age = options.fetch(:age, 0)
-    @name = options.fetch(:name, "")
-  end
-
-  def offer_high_five
-    "High five!"
   end
 
   def set_phase(num)
@@ -30,4 +29,5 @@ class Student
     response += "Now I feel like a genius!"
     response
   end
+  
 end
